+++ pointers/src/main/kotlin/de/md5lukas/waypoints/pointers/variants/CompassPointer.kt
@@ -4,7 +4,6 @@ import de.md5lukas.schedulers.AbstractScheduler
 import de.md5lukas.waypoints.pointers.Pointer
 import de.md5lukas.waypoints.pointers.PointerManager
 import de.md5lukas.waypoints.pointers.Trackable
-import de.md5lukas.waypoints.pointers.config.CompassConfiguration
 import java.util.concurrent.CompletableFuture
 import org.bukkit.Location
 import org.bukkit.entity.Player
@@ -13,8 +12,9 @@ internal class CompassPointer(
     pointerManager: PointerManager,
     player: Player,
     scheduler: AbstractScheduler,
-    private val config: CompassConfiguration
 ) : Pointer(pointerManager, player, scheduler) {
+
+  private val config = pointerManager.configuration.compass
 
   override val interval: Int
     get() = config.interval